+++ Sources/Toki/main.swift
@@ -8,6 +8,10 @@ if let exitCode = StatusCommand.runIfRequested(arguments: CommandLine.arguments)
     exit(exitCode)
 }
 
+if let exitCode = PiCommand.runIfRequested(arguments: CommandLine.arguments) {
+    exit(exitCode)
+}
+
 let app = NSApplication.shared
 let delegate = AppDelegate()
 app.delegate = delegate